Program can now handle inputs like a1 and take them as A1 instead
Added an "N" and an "R" choice for players. N will close the program and R will show the rules again.

Players can only enter Y, N, or R and nothing else after a game is complete.

Fixed IntroMessage() to tell player that a board can only be reset after a game is complete